Changing scheduler priority queue logic
The previous implementation continue the loop over the process table,
after a process switch, without taking in consideration the process
priority, e.g. a process with priority 1 was found in position 5, in the
next search, process in postions [0, 4] could be runnable and have
priority 2, however they won't be considered in the next search.